requirements.txt
jinja2
bullet
pyyaml
